requirements.txt
# native dependencies
Pillow==11.3.0
lxml==5.2.1
psycopg2==2.9.9
Django==4.2.23

# Other
beautifulsoup4==4.12.3
hyperlink==21.0.0
urllib3==1.26.18
Paver==1.3.4
python-slugify==8.0.4
decorator==5.1.1
celery==5.4.0
tqdm==4.66.4
Deprecated==1.2.14
jsonschema==4.22.0
schema==0.7.7
rdflib==6.3.2
smart_open==7.0.4
PyMuPDF==1.26.3
defusedxml==0.7.1
zipstream-ng==1.8.0
reportlab
Pillow

# Django Apps
django-allauth==0.63.6
django-appconf==1.0.6
django-celery-results==2.5.1
django-filter==24.2
django-imagekit==5.0.0
django-taggit==5.0.1
django-markdownify==0.9.5
django-mptt==0.16.0
django-modeltranslation>=0.11,<0.19.0
django-treebeard==4.7.1
django-guardian<2.4.1
django-downloadview==2.3.0
django-polymorphic==3.1.0
django-tastypie<0.15.0
django-tinymce==3.7.1
django-grappelli==4.0.1
django-uuid-upload-path==1.0.0
django-widget-tweaks==1.5.0
django-sequences==3.0
oauthlib==3.3.1
pyjwt==2.8.0

# geopython dependencies
git+https://github.com/geopython/pycsw.git@3.0.0-beta1#egg=pycsw
pyproj<3.7.0
OWSLib==0.31.0
SQLAlchemy==2.0.30 # required by PyCSW
Shapely==1.8.5.post1
mercantile==1.2.1
numpy==1.26.*
Jinja2==3.1.4

# # Apps with packages provided in GeoNode's PPA on Launchpad.

# Django Apps
dj-database-url==2.1.0
dj-pagination==2.5.0
django-select2==8.1.2
django-floppyforms<1.10.0
django-forms-bootstrap<=3.1.0
django-autocomplete-light==3.11.0
django-invitations<2.1.1
django-recaptcha==4.0.0

# REST
djangorestframework==3.14.0
djangorestframework-gis==1.2.0
djangorestframework-guardian==0.3.0
drf-extensions==0.7.1
drf-spectacular==0.27.2
dynamic-rest==2.3.0

geonode-pinax-notifications==6.0.0.2

# GeoNode org maintained apps.
# django-geonode-mapstore-client==4.0.5
git+https://github.com/GeoNode/geonode-mapstore-client.git@master#egg=django_geonode_mapstore_client
django-avatar==8.0.0
geonode-oauth-toolkit==2.2.2.2
geonode-announcements==2.0.2.2
django-activity-stream==2.0.0
gn-arcrest==10.5.6
geoserver-restconfig~=2.0.14
gn-gsimporter==2.0.4
gisdata==0.5.4

# importer dependencies
setuptools>=59
gdal<=3.4.3
pdok-geopackage-validator==0.8.5
geonode-django-dynamic-model==0.4.0


# datetimepicker widget
django-bootstrap3-datetimepicker-2==2.8.3

# storage manager dependencies
django-storages==1.14.3
dropbox==11.36.2
google-cloud-storage==2.16.0
boto3==1.39.3

# Django Caches
whitenoise==6.6.0

# Contribs
xmltodict<0.13.1
inflection>=0.4.0
mock<6.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.4
timeout-decorator==0.5.0
pylibmc==1.6.3
sherlock==0.4.1

psutil==5.9.8
django-cors-headers==4.3.1
django-user-agents==0.4.0

# production
uWSGI==2.0.30
docker==7.0.0
invoke==2.2.0

# tests
coverage==7.5.1
requests-toolbelt==1.0.0
flake8==7.3.0
black==24.4.2
pytest==8.2.0
pytest-bdd==7.1.2
splinter==0.21.0
pytest-splinter==3.3.2
pytest-django==4.8.0
pixelmatch==0.3.0
factory-boy==3.3.0
flaky==3.8.1
selenium>=4.15.1,<5.0.0
selenium-requests==2.0.4
webdriver_manager==4.0.1

# Security and audit
cryptography==45.0.5
jwcrypto>=1.5.6
nh3==0.2.17